+++ tests/CVLB/Svc/Api/Services/Logging/LoggingTest.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace CVLB\Svc\Api\Services;
+namespace CVLB\Svc\Api\Services\Logging;
 
 use CVLB\Svc\Api\AuthService;
 use CVLB\Svc\Api\HttpClient\Message\ResponseMediator;
 use CVLB\Svc\Api\Sdk;
 use Http\Client\Common\HttpMethodsClientInterface;
-use Http\Mock\Client;
 use Mockery;
+use PHPUnit\Framework\MockObject\MockMethod;
 use PHPUnit\Framework\TestCase;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\StreamInterface;
@@ -36,7 +36,7 @@ class LoggingTest extends TestCase
         $mockSdk = Mockery::mock(Sdk::class);
         $mockSdk->shouldReceive('getHttpClient')
             ->andReturn($mockHttpMethodsClientInterface);
-        $mockSdk->shouldReceive('getAUth')
+        $mockSdk->shouldReceive('getAuth')
             ->andReturn($mockAuthService);
 
         unset($_SERVER['SERVER_ADDR']);
@@ -48,12 +48,10 @@ class LoggingTest extends TestCase
 
     public function testPutException()
     {
-        //$this->expectException(\Exception::class);
-
+        $errorMessage = 'Test Error';
         $_ENV['APP_ENV'] = 'local';
         $mockStream = Mockery::mock(StreamInterface::class);
-        $mockStream->shouldReceive('getContents')
-            ->andReturn(json_encode(['OK']));
+        $mockStream->shouldReceive('getContents')->andThrow(new \Exception($errorMessage));
 
         $mockResponseInterface = Mockery::mock(ResponseInterface::class);
         $mockResponseInterface->shouldReceive('getBody')
@@ -74,7 +72,7 @@ class LoggingTest extends TestCase
         $mockSdk = Mockery::mock(Sdk::class);
         $mockSdk->shouldReceive('getHttpClient')
             ->andReturn($mockHttpMethodsClientInterface);
-        $mockSdk->shouldReceive('getAUth')
+        $mockSdk->shouldReceive('getAuth')
             ->andReturn($mockAuthService);
 
 
@@ -83,5 +81,7 @@ class LoggingTest extends TestCase
         $response = $logging->put('string', ['test'=>'data'], 123);
 
         $this->assertIsArray($response);
+        $this->assertEquals(false, $response['success']);
+        $this->assertEquals(500, $response['code']);
     }
 }