+++ tests/CVLB/Svc/Api/Services/Notify/Protocols/Email/EmailTest.php
@@ -1,0 +1,64 @@
+<?php
+
+namespace CVLB\Svc\Api\Services\Notify\Protocols\Email;
+
+use CVLB\Svc\Api\AuthService;
+use CVLB\Svc\Api\Sdk;
+use CVLB\Svc\Api\Services\Notify\Notify;
+use CVLB\Svc\Api\Services\Notify\Protocols\Chat\Chat;
+use CVLB\Svc\Api\Services\Notify\Protocols\Chat\SlackService;
+use Http\Client\Common\HttpMethodsClientInterface;
+use Mockery;
+use PHPUnit\Framework\TestCase;
+use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\StreamInterface;
+
+class EmailTest extends TestCase
+{
+    public function getNotify(): Notify
+    {
+        $_ENV['APP_ENV'] = 'local';
+        $mockStream = Mockery::mock(StreamInterface::class);
+        $mockStream->shouldReceive('getContents')->andReturn(json_encode(['OK']));
+
+        $mockResponseInterface = Mockery::mock(ResponseInterface::class);
+        $mockResponseInterface->shouldReceive('getBody')
+            ->andReturn($mockStream);
+
+        $mockAuthService = Mockery::mock(AuthService::class);
+        $mockAuthService->shouldReceive('getClientId')
+            ->andReturn('sdk_phpunit');
+
+        $mockHttpMethodsClientInterface = Mockery::mock(HttpMethodsClientInterface::class);
+        $mockHttpMethodsClientInterface->shouldReceive('post')
+            ->andReturn($mockResponseInterface);
+
+        $mockSdk = Mockery::mock(Sdk::class);
+        $mockSdk->shouldReceive('getHttpClient')
+            ->andReturn($mockHttpMethodsClientInterface);
+        $mockSdk->shouldReceive('getAuth')
+            ->andReturn($mockAuthService);
+
+        $mockSdk->app_name = 'PHP Unit Test';
+
+        return new Notify($mockSdk);
+    }
+
+    public function testSes()
+    {
+        $notify = $this->getNotify();
+        $email = new Email($notify);
+        $service = $email->ses();
+
+        $this->assertInstanceOf(SesService::class, $service);
+    }
+
+    public function testSendpulse()
+    {
+        $notify = $this->getNotify();
+        $email = new Email($notify);
+        $service = $email->sendpulse();
+
+        $this->assertInstanceOf(SendPulseService::class, $service);
+    }
+}